SDBELGA-279 Published Belga NewsML should contain all objects in the update chain
